Refactor forms as well
