+++ konfera/event/forms.py
@@ -1,1 +1,20 @@
-forms.py
+from django import forms
+
+from konfera.models import Speaker, Talk, Ticket
+
+
+
+class SpeakerForm(forms.ModelForm):
+
+    class Meta:
+        model = Speaker
+        exclude = ['sponsor']
+
+
+class TalkForm(forms.ModelForm):
+
+    class Meta:
+        model = Talk
+        exclude = ['status', 'primary_speaker', 'secondary_speaker', 'event']
+
+
